+++ src/Graphics/Model/mdlFile.c
@@ -620,7 +620,7 @@ void FUN_0033e700(int param_1);
 u32 FUN_0033e7f0(int param_1);
 void FUN_0033e890(int param_1);
 void FUN_0033e8e0(int param_1);
-void FUN_0033f2e0(u64 param_1);
+void FUN_0033f2e0(int param_1);
 void FUN_0033f4a0(int param_1);
 void FUN_0033f590(int param_1,int param_2);
 u64 FUN_0033f6a0(u64 param_1,long param_2);
@@ -35701,7 +35701,7 @@ void FUN_0033e8e0(int param_1)
 // FUN_0033F2E0 NONMATCHING
 
 
-void FUN_0033f2e0(u64 param_1)
+void FUN_0033f2e0(int param_1)
